Cmake update (#845)
* some more work on modernizing the cmake files with best practices

* change cmake variables and update all the CI systems to use them

* split out the versioning string into a separate function

* add a couple layers of error checking in the version_describe.cmake

* move jsoncppobj_lib sources into bool generate expression in library add call instead of separately

* make some fixes to the HELICSConfig file

* update the package tests to use the correct variable

* make some adjustments to the cmake for adding libzmq and an ability to use the src directory as the base build

* add separate cmakeLists heads to benchmarks and tests

* change the linked library name for units.

* use generator expressions on the compiler flags so they aren't duplicated in the file

* more compiler flag updates and fixes to the CMakeLists for helics-shared

* update the aliases in some of the targets

* only make jsoncpp alias is the target exists

* fix a few new warnings from visual studio

* some more warning fixes for windows(and a couple small bug fixes related to the warnings)

* add the extra library folder to the helics-shared as well

* update copyright on all cmake files in prep for formatting

* change the targets in cmakeLists to add HELICS::compile_flags_target

* reupdate concurrency library and add include paths to shared library targets

* try to clear up a bunch of bool conversion warnings

* fix type conversion bug from new code

* updates to cppzmq from upstream to clear up some warnings,  this results in a lot of deprecated functions that we needed to update in the cores and tests.  The warnings were also indicative of potential abort conditions from noexcept mismatch that should be cleared up in the new version.

* fix a bug in one of the changes to zmq.hpp

* put the string getsockopt funciton in the correct location

* a few more tweaks to the zmq.hpp and the cmake options

* another fix for zmq.hpp and C++17

* add more consistent bool conversions across helics and make zmq consistent with upstream repo

* acknowledge and use BUILD_SHARED_LIBS flag

* fix a few warnings and try a slightly different condition on zmq.hpp

* remove barrier code as it is now fully in concurrency library,  update concurrency library with some warning fixes,  remove some message statements from cmake code.

* enable /W4 warnings on visual studio.  clear up some of the new warnings.  move some more options to be HELICS_XXXX, start adding the docs of the options.  Make sure the code uses the new option names instead of the old ones.

* update some of the documentation, finish documenting CMAKE options available in HELICS.

* some minor updates to zmq and helics_cmake_option.md

* fix a few static analysis issues from Visual stdio 2019

* make sure the options used in the CI tests are valid

* run cmake-format on all CMakeLists.txt files

* format config cmake files and use a little better practices for some variable to reduce cmake code

* add some additional macros to extraMacros.cmake final formatting.

* update some warnings for google benchmark

+++ src/helics/core/helics-time.hpp
@@ -33,7 +33,7 @@ constexpr Time timeEpsilon = Time::epsilon ();
 /** definition of the smallest negative increment of time*/
 constexpr Time negEpsilon = -Time::epsilon ();
 /** user defined literal for a time variable*/
-constexpr Time operator"" _t (long double val) { return Time (val); }  // NOLINT
+constexpr Time operator"" _t (long double val) { return Time (static_cast<double> (val)); }  // NOLINT
 
 /** simple structure with the time and completion marker for iterations or dense time steps*/
 struct iteration_time
